Moved styled toolbar into separate file
