+++ src/components/views/MenuResourceSettingsView.tsx
@@ -25,6 +25,7 @@ import VisibleWithRole from "../generic/VisibleWithRole";
 import { getLocalizedTypeString } from "../../commons/resourceTypeHelper";
 import { ErrorContext } from "../containers/ErrorHandler";
 import { resolveChildResourceTypes } from "../../commons/resourceTypeHelper";
+import StyledMTableToolbar from "../../styles/generic/styled-mtable-toolbar";
 
 /**
  * Component props
@@ -410,14 +411,6 @@ class MenuResourceSettingsView extends React.Component<Props, State> {
   private renderStyleTable = () => {
     const { resourceData } = this.state;
 
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
-
     return (
       <MaterialTable
         icons={{
@@ -520,14 +513,6 @@ class MenuResourceSettingsView extends React.Component<Props, State> {
    */
   private renderPropertiesTable = () => {
     const { resourceData } = this.state;
-
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
 
     return (
       <MaterialTable
