+++ src/components/views/PDFResourceSettingsView.tsx
@@ -18,6 +18,7 @@ import VisibleWithRole from "../generic/VisibleWithRole";
 import { ErrorContext } from "../containers/ErrorHandler";
 import PDFPreview from "../generic/PDFPreview";
 import theme from "../../styles/theme";
+import StyledMTableToolbar from "../../styles/generic/styled-mtable-toolbar";
 
 /**
  * Component props
@@ -237,14 +238,6 @@ class PDFResourceSettingsView extends React.Component<Props, State> {
   private renderStyleTable = () => {
     const { resourceData } = this.state;
 
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
-
     return (
       <MaterialTable
         icons={{
@@ -347,14 +340,6 @@ class PDFResourceSettingsView extends React.Component<Props, State> {
    */
   private renderPropertiesTable = () => {
     const { resourceData } = this.state;
-
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
 
     return (
       <MaterialTable
