+++ src/components/views/ResourceSettingsView.tsx
@@ -20,6 +20,7 @@ import { AuthState, ErrorContextType } from "../../types";
 import { ErrorContext } from "../containers/ErrorHandler";
 import { connect } from "react-redux";
 import { ReduxState } from "../../store";
+import StyledMTableToolbar from "../../styles/generic/styled-mtable-toolbar";
 
 /**
  * Component props
@@ -266,14 +267,6 @@ class ResourceSettingsView extends React.Component<Props, State> {
   private renderStyleTable = () => {
     const { resourceData } = this.state;
 
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
-
     return (
       <MaterialTable
         icons={{
@@ -367,14 +360,6 @@ class ResourceSettingsView extends React.Component<Props, State> {
    */
   private renderPropertiesTable = () => {
     const { resourceData } = this.state;
-
-    const StyledMTableToolbar = withStyles({
-      root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2),
-        borderBottom: "1px solid rgba(0,0,0,0.1)"
-      },
-    })(MTableToolbar);
 
     return (
       <MaterialTable
